backend/subscriptions/views.py
from django.shortcuts import render

# Create your views here.
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from drf_spectacular.utils import extend_schema

@api_view(["GET"])
def health(request):
    return Response({"status": "ok"})

# The first line is just for documentation
@extend_schema(summary="Who am I", tags=["Auth"])
@api_view(["GET"])
@permission_classes([IsAuthenticated])
def me(request):
    user = request.user
    return Response({"id": user.id, "username": user.username, "email": user.email})

